Read initial EMMS stream definitions from streams file
In EMMS 6.0, streams are defined in a native playlist located in
`emms-streams-file'.  When no streams are definied initially in
`helm-emms-streams-list', let `helm-emms' read the list from `emms-streams-file'
by default, and fall back to `emms-streams-built-in-list' if that is nil.  By
this, user defined streams are shown by `helm-emms' instead of the builtin ones.