Highlight currently played file when using custom track descriptions
When `helm-emms-use-track-description-function` is non-nil, candiates for
`helm-source-emms-files` have a car formatted by
`emms-track-description-function`.  To properly highlight the currently played
track in this case, `helm-emms-files-modifier` should honor the custom track
formatting first, such that the following `string-match-p` can succeed even if
the currently played track has an info-title.  This can be achieved by switching
the order of tests when defining the `playing` variable in
`helm-emms-files-modifier`.  Moreover, `emms-playlist-current-selected-track`
returns a track object, so the `stringp` test can never succed and is thus
removed.